Initial commit
Note to professor:

The TechNOVR team unfortunately did not utilize GitHub throughout the project. To avoid downloading and re-uploading to PlayCanvas, we used PlayCanvas's internal VCS which the commit history could not be exported to GitHub. It is possible to export each commit individually and recreate the structure in GitHub, but that would involve downloading each commit form PlayCanvas, copying the changed files, and making sure that git was tracking each files changes properly and not just marking each modified file as entirely changed. If there is a need, we can go through this process to show our commit history, although the history is quite slim as it is.